+++ CHANGELOG.md
@@ -6,6 +6,7 @@ master
  - PhotoUrl - gdy puste photoId zwraca link do zaślepiki noimg<size>.png
  - ProductsRepository - dodanie pola 'click_url' do 'fieldsForProductBox'
  - Offer - dodanie pola join id
+ - OffersRepository - funkcja pobierania oferty po joinId
 
 
 v1.3.1